Bugfix: Special handling for multiple "totalCases" entries in the same Gaius API response
This fix addresses an issue that cropped up with the Gaius API.  Per the logic in GaiusAPIService, it was required to remove the "totalCases" entries to facilitate a key-value abstraction.  Expected response looks like:

```json
{
    "1": {
        "userid": "0000000000000000000",
        "type": "ban",
        "time": null,
        "modid": "0000000000000000000",
        "reason": "Likes pineapple on pizza",
        "date": 0,
        "casenum": 1
    },
    "totalCases": "1"
}
```

However, as of November 27th, 2023, there was a change to the Gaius API that resulted in two totalCases entries returned, like so:

```json
{
    "1": {
        "userid": "0000000000000000000",
        "type": "ban",
        "time": null,
        "modid": "0000000000000000000",
        "reason": "Likes pineapple on pizza",
        "date": 0,
        "casenum": 1
    },
    "totalcases": 3,
    "totalCases": "1"
}
```

Worse still, this new entry "totalcases" does not match the value of "totalCases", often sitting idly at 3.  To address this locally, I have opted to change the logic of the code to simply remove any lines containing a JSON key with any casing of "totalcases".